# api/dto/loai_doc_gia.go
package dto

type PostLoaiDocGiaDto struct {
	TenLoaiDocGia string `json:"tenLoaiDocGia" binding:"required"`
}

type PutLoaiDocGiaDto struct {
	TenLoaiDocGia string `json:"tenLoaiDocGia" binding:"required"`
}