# docs/api_doc.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Library Management Service API",
    "description": "Library Management Service API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost"
    }
  ],
  "paths": {
...
    "/loaidocgia": {
      "get": {
        "tags": [
          "Loai doc gia"
        ],
        "description": "Get danh sach loai doc gia",
        "responses": {
          "200": {
            "description": "Get danh sach loai doc gia successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LoaiDocGiaPresenter"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": [
          "Loai doc gia"
        ],
        "description": "Create loai doc gia",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostLoaiDocGiaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create loai doc gia successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoaiDocGiaPresenter"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/loaidocgia/{maLoaiDocGia}": {
      "get": {
        "tags": [
          "Loai doc gia"
        ],
        "description": "Get loai doc gia by ma loai doc gia",
        "parameters": [
          {
            "in": "path",
            "name": "maLoaiDocGia",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Get loai doc gia successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoaiDocGiaPresenter"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "put": {
        "tags": [
          "Loai doc gia"
        ],
        "description": "Update loai doc gia by ma loai doc gia",
        "parameters": [
          {
            "in": "path",
            "name": "maLoaiDocGia",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutLoaiDocGiaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update loai doc gia successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoaiDocGiaPresenter"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "tags": [
          "Loai doc gia"
        ],
        "description": "Delete loai doc gia by ma loai doc gia",
        "parameters": [
          {
            "in": "path",
            "name": "maLoaiDocGia",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Delete loai doc gia successful"
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "409": {
            "$ref": "#/components/responses/ConflictError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/phieunhap": {
      "get": {
        "summary": "Get danh sach phieu nhap",
        "tags": [
          "Phieu nhap"
        ],
        "responses": {
          "200": {
            "description": "Get danh sach phieu nhap successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PhieuNhapPresenter"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "summary": "Create phieu nhap",
        "tags": [
          "Phieu nhap"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PhieuNhapDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create dau sach successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhieuNhapPresenter"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/phieunhap/{maPhieuNhap}": {
      "get": {
        "summary": "Get phieu nhap",
        "tags": [
          "Phieu nhap"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "maPhieuNhap",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Get phieu nhap successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhieuNhapPresenter"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "patch": {
        "summary": "Update phieu nhap",
        "tags": [
          "Phieu nhap"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "maPhieuNhap",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PhieuNhapDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Update phieu nhap successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhieuNhapPresenter"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "summary": "Remove phieu nhap",
        "tags": [
          "Phieu nhap"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "maPhieuNhap",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Remove phieu nhap successful"
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "summary": "Add chi tiet phieu nhap",
        "tags": [
          "Phieu nhap"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "maPhieuNhap",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CtPhieuNhapDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Add chi tiet phieu nhap successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CtPhieuNhapPresenter"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/phieunhap/{maPhieuNhap}/{maSach}": {
      "delete": {
        "summary": "Remove chi tiet phieu nhap",
        "tags": [
          "Phieu nhap"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "maPhieuNhap",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "in": "path",
            "name": "maSach",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Remove chi tiet phieu nhap successful"
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "PostLoaiDocGiaDto": {
        "type": "object",
        "properties": {
          "tenLoaiDocGia": {
            "type": "string"
          }
        }
      },
      "PutLoaiDocGiaDto": {
        "type": "object",
        "properties": {
          "tenLoaiDocGia": {
            "type": "string"
          }
        }
      },
      "LoaiDocGiaPresenter": {
        "type": "object",
        "properties": {
          "maLoaiDocGia": {
            "type": "string"
          },
          "tenLoaiDocGia": {
            "type": "string"
          }
        }
      },
...
      "TheLoaiPresenter": {
        "type": "object",
        "properties": {
          "maTheLoai": {
            "type": "string"
          },
          "tenTheLoai": {
            "type": "string"
          }
        }
      },
...
      "TacGiaPresenter": {
        "type": "object",
        "properties": {
          "maTacGia": {
            "type": "string"
          },
          "tenTacGia": {
            "type": "string"
          }
        }
      },
...
      "DauSachPresenter": {
        "type": "object",
        "properties": {
          "maDauSach": {
            "type": "string"
          },
          "tenDauSach": {
            "type": "string"
          },
          "theLoai": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TheLoaiPresenter"
            }
          },
          "tacGia": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TacGiaPresenter"
            }
          }
        }
      },
      "PhieuNhapPresenter": {
        "type": "object",
        "properties": {
          "maPhieuNhap": {
            "type": "string"
          },
          "ngayLap": {
            "type": "string",
            "format": "date-time"
          },
          "tongTien": {
            "type": "integer"
          },
          "ctPhieuNhap": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CtPhieuNhapPresenter"
            }
          }
        }
      },
      "CtPhieuNhapPresenter": {
        "type": "object",
        "properties": {
          "sach": {
            "$ref": "#/components/schemas/SachPresenter"
          },
          "donGia": {
            "type": "integer"
          }
        }
      },
      "SachPresenter": {
        "type": "object",
        "properties": {
          "maSach": {
            "type": "string"
          },
          "dauSach": {
            "$ref": "#/components/schemas/DauSachPresenter"
          },
          "nhaXuatBan": {
            "type": "string"
          },
          "triGia": {
            "type": "integer"
          },
          "namXuatBan": {
            "type": "integer"
          },
          "tinhTrang": {
            "type": "boolean"
          }
        }
      },
      "PhieuNhapDto": {
        "type": "object",
        "properties": {
          "ngayLap": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CtPhieuNhapDto": {
        "type": "object",
        "properties": {
          "maDauSach": {
            "type": "string"
          },
          "nhaXuatBan": {
            "type": "string"
          },
          "triGia": {
            "type": "integer"
          },
          "namXuatBan": {
            "type": "integer"
          },
          "tinhTrang": {
            "type": "boolean"
          },
          "donGia": {
            "type": "integer"
          }
        }
      }
    },
    "responses": {
      "BadRequestError": {
        "description": "provide bad request"
      },
      "ForbiddenError": {
        "description": "do not have permission to view this resource"
      },
      "NotFoundError": {
        "description": "resource not found"
      },
      "UnauthorizedError": {
        "description": "token is missing"
      },
      "ConflictError": {
        "description": "conflict with state of server, maybe due to database constraint reason,.."
      },
      "InternalServerError": {
        "description": "internal server error"
      }
    }
  }
}